MdeModulePkg/SmartCardReader: Add files from the libccid project
libccid project is hosted at http://pcsclite.alioth.debian.org/ccid.html

This code is shared with the libccid project. In order to easily port
changes between libccid and UEFI projects I decided to keep the coding
style as it was and not strictly comply to the UEFI coding style.

Contributed-under: GNU LGPL v2.1+
Signed-off-by: Ludovic Rousseau <ludovic.rousseau@gmail.com>